Refactor the authentication parser
